Added in the start of the graphing feature
two functions updater() and counter(). updater() adds onto an array the current bearing and can take a reverse parameter which flips the bearings around to account for times when the rover backs up. counter() is used for debug purposes and just checks to see if the array is made up of numbers other than 0.0. I still need to account for times when the rover may be stuck by adding in a movement checker.